move mocks and utils around; fix error logging
