+++ src/infrastructure/jira/index.ts
@@ -1,2 +1,3 @@
 export * from './jira-service';
+export * from './utils';
 export * from './errors';
