+++ src/infrastructure/jira/utils.ts
