+++ src/usecases/disassociate-entity-use-case.ts
@@ -1,11 +1,10 @@
 import type { AtlassianEntity } from './types';
-import { buildIssueUrl } from './utils';
 
 import { JIRA_ISSUE_ATI } from '../common/constants';
 import type { AtlassianDesign, ConnectInstallation } from '../domain/entities';
 import { AtlassianAssociation } from '../domain/entities';
 import { figmaService } from '../infrastructure/figma';
-import { jiraService } from '../infrastructure/jira';
+import { buildIssueUrl, jiraService } from '../infrastructure/jira';
 
 export type DisassociateEntityUseCaseParams = {
 	readonly entity: {
